% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{Analysis.DISEXP}
\alias{Analysis.DISEXP}
\title{Perform correlation analysis based on RNA distance and Gene expression}
\usage{
Analysis.DISEXP(
  dis.name,
  dis.distance,
  exp.tumor,
  exp.sample,
  method = "linear",
  showPlot = FALSE
)
}
\arguments{
\item{dis.name}{Set of name of RNA distance}

\item{dis.distance}{Set of RNA distance between mutate and original data}

\item{exp.tumor}{Set of reads from gene expression from tumor samples}

\item{exp.sample}{Set of reads from gene expression from normal samples (usually blood sample)}

\item{method}{Selection of linear or gaussian log link function for regression (linear or log)}

\item{showPlot}{TRUE and FALSE variable if TRUE the output image will be shown on the run}
}
\value{
Returns an S3 object of class DISEXP with results.
list of output stats from the model
\itemize{
\item stats
\itemize{
  \item Correlation - Beta value of the regression model based on data
  \item PValue - P value calculated associated with the correlation
}
Show of four graph for validation
\item plots
\itemize{
  \item ScatterPlot plots of gene expression data on RNA distance data with fitted line
  \item GeneExpressionBoxPlot Boxplot checking outliars on gene expression
  \item RNADistanceBoxPlot Boxplot checking outliars on distance
  \item RNADistanceDensityPlot  Density plot on distance by different RNA
  \item GeneExpressionDensityPlot Density plot on gene expression by different RNA
}
}
}
\description{
Analysis.DISEXP is a complete analysis based on user selection of
linear or log regression. The gene expression is calculated as the
absolute differences between sampled and normal gene expression data.
Analysis also export sets of graphs to facilitate in model selection
and analysis result validation.
}
\examples{
disexp <- Analysis.DISEXP(
             dis.name = c("hsa-let-7a-1", "hsa-let-7a-1",
             "hsa-let-7a-3", "hsa-let-7a-3", "hsa-let-7a-3"),
             dis.distance = as.integer(c(10, 35, 91, 100, 92)),
             exp.tumor = c(98691, 49201, 57540, 148702, 97721),
             exp.sample = c(23495, 23310, 13274, 19337, 14389))

}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
