% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{bed2df}
\alias{bed2df}
\title{Return a dataframe object with RNA location information}
\usage{
bed2df(filepath)
}
\arguments{
\item{filepath}{The path directed to bed file on current working directory}
}
\value{
Return a data.frame object contain bed document loaded
}
\description{
Perform bed file retrieving using result stored
in standardized list for easy usage.
}
\examples{
bed2df("./test.bed")

}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
