% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{fasta2df}
\alias{fasta2df}
\title{Return a dataframe object with fasta information}
\usage{
fasta2df(filepath)
}
\arguments{
\item{filepath}{The path directed to fasta file on current working directory}
}
\value{
Return a data.frame object contain fasta document loaded
\itemize{
  \item NAME - Corresponding of the RNA sequence
  \item SEQ - The original RNA sequence
}
}
\description{
Perform fasta file retrieving using readRNAStringSet function from Biostrings, result stored
in standarlize data.frame for easy usage.
}
\examples{
\dontrun{fasta2df("./test.fasta")}

}
\references{
H. Pagès, P. Aboyoun, R. Gentleman and S. DebRoy (2020). Biostrings: Efficient manipulation of biological strings. R package version 2.56.0.
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
