% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predictDistance_gsc}
\alias{predictDistance_gsc}
\title{(Internal Function) Return list of calculated RNA distance using gscVisualizer}
\usage{
predictDistance_gsc(name = c(), struct.ori = c(), struct.alt = c())
}
\arguments{
\item{name}{List consisted with name of the RNA sequence in seq}

\item{struct.ori}{List consisted with RNA orginal sequence}

\item{struct.alt}{List consisted with RNA alternative sequence}
}
\value{
Returns a list of RNA distance, ordered according to the input name sequence
}
\description{
Perform RNA distance calculation based on imported file or
input of RNA original sequence and alternative sequence. The RNA
distance calculation is based on the gscVisualizer R package
algorithm by Zhiwen Tan, a BCB student from UofT.
}
\references{
{Z Tan (2020). gscVisualizer: Visualize The Difference among Gene Sequences. R package version 0.1.0.}
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
