% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predictDistance_RNADis}
\alias{predictDistance_RNADis}
\title{(Internal Function) Return list of calculated RNA distance using RNADistance}
\usage{
predictDistance_RNADis(
  executable.path = "",
  name = c(),
  struct.ori = c(),
  struct.alt = c()
)
}
\arguments{
\item{executable.path}{The path referred to RNADistance executable (.../RNADistance) used by program}

\item{name}{List consisted with name of the RNA sequence in seq}

\item{struct.ori}{List consisted with RNA orginal sequence}

\item{struct.alt}{List consisted with RNA alternative sequence}
}
\value{
Returns a list of RNA distance, ordered according to the input name sequence
}
\description{
Perform RNA distance calculation based on imported file or
input of RNA orginal sequence and alternative sequence. The RNA
distance calculation is based on the RNAdistance command line
algorithm by Walter Fontana, Ivo L Hofacker, Peter F Stadler.
RNA distance calculation is based on -XF parameter, which
indicates single alignment among each comparisons.
MacOS or Unix user can ignore executable.path, when RNADistance
is already installed in the system tools.
}
\examples{
dot <- predictDistance_RNADis(name = c("hsa-let-7b", "hsa-let-7a-2")
, struct.ori = c("(((((.((((((((((((((((((((((((((((((.....))).)))).))).....)))))))))))))))))))))))))",
          "(((((.((((((((((((((((((((((((((((((.....))).)))).))).....)))))))))))))))))))))))))")
, struct.alt = c("(((((.(((((((((((((((((((..(((((((((.....)))))).).....))...))))))))))))))))))))))))",
           "(((((.((((((((.(((((((((((((((((((((.....)))))).).))).....))))))))))).)))))))))))))"))

}
\references{
Lorenz, Ronny and Bernhart, Stephan H. and Höner zu Siederdissen, Christian and Tafer, Hakim and Flamm, Christoph and Stadler, Peter F. and Hofacker, Ivo L.
ViennaRNA Package 2.0 Algorithms for Molecular Biology, 6:1 26, 2011, doi:10.1186/1748-7188-6-26
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
