% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{predictStructure}
\alias{predictStructure}
\title{Return list of predicted secondary structure}
\usage{
predictStructure(
  executable.path = "",
  fasta.file = "",
  rna.name = c(),
  rna.seq = c()
)
}
\arguments{
\item{executable.path}{The path referred to RNAStructure executable (.../RNAStructure/exe) used by program}

\item{fasta.file}{The file path with mature sequences in fasta format}

\item{rna.name}{List consisted with name of the RNA sequence in seq}

\item{rna.seq}{List consisted with RNA sequence}
}
\value{
Returns a list of dot bracket form, ordered according to the input name sequence
}
\description{
Perform RNA structural prediction based on fasta file or
input of RNA name and sequence. Secondary structure prediction
algorithm based on command line program RNAStructure from
Mathews Lab, program require current path to the executable,
MacOS or Unix based user installed entire package is allowed to
omit.
}
\examples{
\dontrun{dot <- predictStructure(executable.path = "./exe"
                                 , rna.name = c("hsa-let-7b", "hsa-let-7a-2")
                                 , rna.seq = c("UGGGAUGAGGUGGAUGUCUUUCCUA", "GAUAACUAUACAAUCUAC"))}

}
\references{
Duan, S., Mathews, D.H. and Turner, D.H. (2006).
Interpreting oligonucleotide microarray data to determine RNA secondary structure: application to the 3' end of Bombyx mori R2 RNA.
Biochemistry, 45:9819-9832.

Wuchty, S., Fontana, W., Hofacker, I.L. and Schuster, P. (1999).
Complete suboptimal folding of RNA and the stability of secondary structures.
Biopolymers, 49:145-165.
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
