% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{RNA.validate}
\alias{RNA.validate}
\title{Apply DNA mutation to RNA and validate the file}
\usage{
RNA.validate(fasta, vcf, bed)
}
\arguments{
\item{fasta}{RNA sequence to be matched read using fasta2df}

\item{vcf}{mutation information read using vcf2df function}

\item{bed}{bed information read using bed2df function}
}
\value{
Return the RNA.mutated data frame that includes all information
about the RNA and its information
\itemize{
  \item NAME - Corresponding of the RNA sequence
  \item MATCH - Boolean value indicating whether the mutation can be applied,
        FALSE value indicate the sequence given has inconsistent bases comparing to the
        reference base in VCF file
  \item STAPOS - The start location of the sequence
  \item ENDPOS - The end location of the sequence
  \item DIR - The direction of the sequence
  \item SEQ - The original RNA sequence
  \item CUR.REF - The current RNA base(s) on the mutation point
  \item MUT.REF - The reference base(s) on the mutation point
  \item MUT.ALT - The alternative base(s) appeared in during mutation
  \item MUT.SEQ - The mutated RNA sequence when MATCH is TRUE, otherwise empty string
  \item MUT.POS - The start position of the mutation
}
}
\description{
Calculate the mutated series of RNA from RNA fasta, bed file
, and DNA vcf file, both should be result file from vcf2df
and fasta2df file, function gives validation based on fasta
and vcf comparisons, user should user validation result to make
sure their vcf and fasta file are aligned under same version.
Algorithm inspired by the mRNA mutation function, a utility function
created by Sijie Xu and Zhiwen Tan, the updated method here is a more generalized
and input validated method, more checking are embeded to insure the
mutation is applied correctly.
}
\examples{
\dontrun{mutated <- RNA.validate(
              fasta = fasta2df("./test.fasta"),
              vcf = vcf2df("./test.vcf",
              bed = bed2df("./test.bed"))}

}
\references{
Xu, S. Tan, Z., BCB430 "Analysis.zip/File Validation.rmd" <https://github.com/Deemolotus/BCB330Y-and-BCB430Y>
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
