% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{runRSEAnalysis}
\alias{runRSEAnalysis}
\title{Launch Shiny App For Package rseAnalysis}
\usage{
runRSEAnalysis()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
Analysis.DISEXP is a complete analysis based on user selection of
linear or log regression. The gene expression is calculated as the
absolute differences between sampled and normal gene expression data.
Analysis also export sets of graphs to facilitate in model selection
and analysis result validation.
}
\examples{
\dontrun{
runRSEAnalysis()
}

}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
