% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{vcf2df}
\alias{vcf2df}
\title{Return a dataframe object with gene mutation information}
\usage{
vcf2df(filepath)
}
\arguments{
\item{filepath}{The path directed to vcf file on current working directory}
}
\value{
Return a data.frame object contain vcf document loaded
}
\description{
Perform vcf file retrieving using read.vcfR function, result stored
in standarlize data.frame for easy usage.
}
\examples{
vcf2df("./test.vcf")

}
\references{
Knaus BJ, Grünwald NJ (2017). “VCFR: a package to manipulate and visualize variant call format data in R." _Molecular Ecology Resources_, *17*(1), 44-53. ISSN 757,
<URL: http://dx.doi.org/10.1111/1755-0998.12549>.
}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
