% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{Analysis.DISEXP}
\alias{Analysis.DISEXP}
\title{Perform correlation analysis based on RNA distance and Gene expression}
\usage{
Analysis.DISEXP(
  dis.name,
  dis.distance,
  exp.tumor,
  exp.sample,
  method = "linear"
)
}
\arguments{
\item{dis.name}{Set of name of RNA distance}

\item{dis.distance}{Set of RNA distance between mutate and original data}

\item{exp.tumor}{Set of reads from gene expression from tumor samples}

\item{exp.sample}{Set of reads from gene expression from normal samples (usually blood sample)}

\item{method}{Selection of linear or gaussian log link function for regression}
}
\value{
Show of four graph for validation, includes
\itemize{
  \item Scatter plot of gene expression data on RNA distance data
  \item Boxplot of gene expression data by each RNA
  \item Boxplot of RNA distance data by each RNA
  \item Density plot of RNA distance data, labelled by each RNA
}
list of output indicating the corrolation\itemize{
  \item correlation - Beta value of the regression model based on data
  \item p_value - P value calculated associated with the correlation
}
}
\description{
Analysis.DISEXP is a complete analysis based on user selection of
linear or log regression. The gene expression is calculated as the
absolute differences between sampled and normal gene expression data.
Analysis also export sets of graphs to facilitate in model selection
and analysis result validation.
}
\examples{
\dontrun{disexp <- Analysis.DISEXP(
             dis.name = c("hsa-let-7a-1", "hsa-let-7a-1",
             "hsa-let-7a-3", "hsa-let-7a-3", "hsa-let-7a-3"),
             dis.distance = as.integer(c(10, 35, 91, 100, 92, 5)),
             exp.tumor = c(98691, 49201, 57540, 148702, 97721),
             exp.sample = c(23495, 23310, 13274, 19337, 14389))}

}
\author{
Sijie Xu, \email{sijie.xu@mail.utoronto.ca}
}
